+++ weblate/settings_example.py
@@ -577,6 +577,7 @@ CRISPY_TEMPLATE_PACK = 'bootstrap3'
 #     'weblate.trans.checks.chars.EndQuestionCheck',
 #     'weblate.trans.checks.chars.EndExclamationCheck',
 #     'weblate.trans.checks.chars.EndEllipsisCheck',
+#     'weblate.trans.checks.chars.EndSemicolonCheck',
 #     'weblate.trans.checks.chars.MaxLengthCheck',
 #     'weblate.trans.checks.format.PythonFormatCheck',
 #     'weblate.trans.checks.format.PythonBraceFormatCheck',